+++ dune/uggrid/low/heaps.h
@@ -70,11 +70,6 @@ enum HeapType {GENERAL_HEAP,                  /**< Heap with alloc/free mechanis
                SIMPLE_HEAP         /**< Heap with mark/release mechanism*/
 };
 
-enum [[deprecated]] HeapAllocMode
-{FROM_TOP=1,                       /**< Allocate from top of stack      */
- FROM_BOTTOM=2                       /**< Allocate from bottom of stack   */
-};
-
 /****************************************************************************/
 /****************************************************************************/
 /** @name Defines and macros for the virtual heap management                 */
@@ -136,16 +131,6 @@ void         DisposeMem             (HEAP *theHeap, void *buffer);
 INT          MarkTmpMem             (HEAP *theHeap, INT *key);
 void        *GetTmpMem              (HEAP *theHeap, MEM n, INT key);
 INT          ReleaseTmpMem          (HEAP *theHeap, INT key);
-[[deprecated("Mark taking a mode is deprecated")]] inline INT
-             Mark                   (HEAP *theHeap, INT mode, INT *key)
-{
-  return MarkTmpMem(theHeap,key);
-}
-[[deprecated("Release taking a mode is deprecated")]] inline INT
-             Release                (HEAP *theHeap, INT mode, INT key)
-{
-  return ReleaseTmpMem(theHeap,key);
-}
 /* @} */
 
 END_UG_NAMESPACE