+++ CHANGELOG.md
@@ -5,6 +5,20 @@ SPDX-License-Identifier: LGPL-2.1-or-later
 
 # dune-uggrid 2.10 (unreleased)
 
+* Remove deprecated `AllocEnvMemory` and `FreeEnvMemory`. They were
+  wrappers of standard `malloc` and `free`.
+
+* Remove deprecated `Mark` and `Release`. Use `MarkTmpMem` and
+  `ReleaseTmpMem` instead.
+
+* Remove deprecated `HeapAllocMode` and its flags `FROM_BOTTOM` and `FROM_TOP`.
+  They lost any influence, as we use the system heap.
+
+* Remove deprecated `DDD_InfoProcList`, use `DDD_InfoProcListRange` instead.
+
+* Remove deprecated `DDD_IFOneway`, use method overload with context as first
+  argument instead.
+
 # dune-uggrid 2.9 (2022-11-25)
 
 * The `dune-uggrid` module does not set the preprocessor flag `HAVE_UG` anymore.